+++ eras/shelley-ma/test-suite/test/Test/Cardano/Ledger/Mary/Examples/MultiAssets.hs
@@ -95,7 +95,7 @@ pp =
     & ppMinUTxOValueL .~ Coin 100
 
 ledgerEnv :: SlotNo -> LedgerEnv MaryEra
-ledgerEnv s = LedgerEnv s Nothing minBound pp (AccountState (Coin 0) (Coin 0)) False
+ledgerEnv s = LedgerEnv s Nothing minBound pp (AccountState (Coin 0) (Coin 0))
 
 feeEx :: Coin
 feeEx = Coin 3