+++ libs/cardano-ledger-test/src/Test/Cardano/Ledger/Constrained/Preds/Tx.hs
@@ -893,7 +893,7 @@ gone = do
   (PParamsF _ pp) <- monadTyped (findVar (unVar (pparams proof)) env)
   accntState <- monadTyped (runTarget env accountStateT)
   utxo1 <- monadTyped (findVar (unVar (utxo proof)) env)
-  let lenv = LedgerEnv slot Nothing txIx pp accntState False
+  let lenv = LedgerEnv slot Nothing txIx pp accntState
   -- putStrLn (show (pcTx Babbage tx))
   pure $ case applySTSByProof proof (TRC (lenv, ledgerstate, tx)) of
     Right _ledgerState' -> do
@@ -1002,7 +1002,7 @@ oneTest sizes proof = do
   (PParamsF _ pp) <- monadTyped (runTerm env2 (pparams proof))
   accntState <- monadTyped (runTarget env2 accountStateT)
   txIx <- arbitrary
-  let lenv = LedgerEnv slot Nothing txIx pp accntState False
+  let lenv = LedgerEnv slot Nothing txIx pp accntState
   case applySTSByProof proof (TRC (lenv, ledgerstate, tx)) of
     Right ledgerState' -> pure (totalAda ledgerState' === totalAda ledgerstate)
     Left errs -> do